+++ markdown/src/test/java/it/niedermann/android/markdown/markwon/textwatcher/LowerIndentionTextWatcherTest.java
@@ -1,11 +1,33 @@
 package it.niedermann.android.markdown.markwon.textwatcher;
 
+import android.view.KeyEvent;
+import android.widget.EditText;
+
+import androidx.arch.core.executor.testing.InstantTaskExecutorRule;
+import androidx.test.core.app.ApplicationProvider;
+
+import junit.framework.TestCase;
+
+import org.junit.Before;
+import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.robolectric.RobolectricTestRunner;
 
+import it.niedermann.android.markdown.markwon.MarkwonMarkdownEditor;
+
 @RunWith(RobolectricTestRunner.class)
-public class LowerIndentionTextWatcherTest extends AbstractTextWatcherTest {
+public class LowerIndentionTextWatcherTest extends TestCase {
+
+    @Rule
+    public InstantTaskExecutorRule instantTaskExecutorRule = new InstantTaskExecutorRule();
+
+    private EditText editText;
+
+    @Before
+    public void reset() {
+        this.editText = new MarkwonMarkdownEditor(ApplicationProvider.getApplicationContext());
+    }
 
     @Test
     public void shouldLowerIndentionByTwoWhenPressingBackspaceOnAnIndentedList() {
@@ -83,5 +105,16 @@ public class LowerIndentionTextWatcherTest extends AbstractTextWatcherTest {
         this.editText.setText("- [ ] Foo");
         pressBackspace(9);
         assertText("- [ ] Fo", 8);
+    }
+
+    private void assertText(String expected, int cursorPosition) {
+        assertEquals(expected, this.editText.getText().toString());
+        assertEquals(cursorPosition, this.editText.getSelectionStart());
+        assertEquals(cursorPosition, this.editText.getSelectionEnd());
+    }
+
+    private void pressBackspace(int atPosition) {
+        this.editText.setSelection(atPosition);
+        this.editText.dispatchKeyEvent(new KeyEvent(KeyEvent.ACTION_DOWN, KeyEvent.KEYCODE_DEL));
     }
 }
